HTTP/requirements.txt
anyio==4.11.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.38.25
botocore==1.38.25
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.7.1
cryptography==45.0.4
daphne==4.2.0
dj-database-url==2.1.0
Django==5.0.14
django-auditlog==3.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-rest-passwordreset==1.5.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.3.0
geographiclib==2.1
geopy==2.4.1
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
msgpack==1.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
s3transfer==0.13.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
supervisor==4.2.5
tomli==2.2.1
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
whitenoise==6.9.0
xlwt==1.3.0
zope.interface==7.2
